// backend/src/types/generated/public/CampaignPromoCodes.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { MarketingCampaignsId } from './MarketingCampaigns';
import type { PromotionsId } from './Promotions';

/** Identifier type for public.campaign_promo_codes */
export type CampaignPromoCodesId = string & { __brand: 'public.campaign_promo_codes' };

/**
 * Represents the table public.campaign_promo_codes
 * Связь маркетинговых кампаний с промоакциями (M:N)
 */
export default interface CampaignPromoCodes {
  id: CampaignPromoCodesId;

  campaign_id: MarketingCampaignsId;

  promotion_id: PromotionsId;

  created_at: string | null;
}

/**
 * Represents the initializer for the table public.campaign_promo_codes
 * Связь маркетинговых кампаний с промоакциями (M:N)
 */
export interface CampaignPromoCodesInitializer {
  /** Default value: gen_random_uuid() */
  id?: CampaignPromoCodesId;

  campaign_id: MarketingCampaignsId;

  promotion_id: PromotionsId;

  /** Default value: now() */
  created_at?: string | null;
}

/**
 * Represents the mutator for the table public.campaign_promo_codes
 * Связь маркетинговых кампаний с промоакциями (M:N)
 */
export interface CampaignPromoCodesMutator {
  id?: CampaignPromoCodesId;

  campaign_id?: MarketingCampaignsId;

  promotion_id?: PromotionsId;

  created_at?: string | null;
}