// backend/src/types/generated/public/Promotions.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.promotions */
export type PromotionsId = string & { __brand: 'public.promotions' };

/** Represents the table public.promotions */
export default interface Promotions {
  id: PromotionsId;

  slug: string;

  title: string;

  description: string;

  image_url: string | null;

  discount_percent: number | null;

  discount_amount: string | null;

  original_price: string | null;

  promo_price: string | null;

  promo_code: string | null;

  usage_limit: number | null;

  usage_count: number | null;

  service_slug: string | null;

  cta_text: string | null;

  cta_url: string | null;

  conditions: string | null;

  starts_at: string | null;

  ends_at: string | null;

  is_active: boolean | null;

  sort_order: number | null;

  created_at: string | null;

  updated_at: string | null;
}

/** Represents the initializer for the table public.promotions */
export interface PromotionsInitializer {
  /** Default value: gen_random_uuid() */
  id?: PromotionsId;

  slug: string;

  title: string;

  description: string;

  image_url?: string | null;

  discount_percent?: number | null;

  discount_amount?: string | null;

  original_price?: string | null;

  promo_price?: string | null;

  promo_code?: string | null;

  usage_limit?: number | null;

  /** Default value: 0 */
  usage_count?: number | null;

  service_slug?: string | null;

  /** Default value: 'Подробнее'::character varying */
  cta_text?: string | null;

  cta_url?: string | null;

  conditions?: string | null;

  starts_at?: string | null;

  ends_at?: string | null;

  /** Default value: true */
  is_active?: boolean | null;

  /** Default value: 0 */
  sort_order?: number | null;

  /** Default value: now() */
  created_at?: string | null;

  /** Default value: now() */
  updated_at?: string | null;
}

/** Represents the mutator for the table public.promotions */
export interface PromotionsMutator {
  id?: PromotionsId;

  slug?: string;

  title?: string;

  description?: string;

  image_url?: string | null;

  discount_percent?: number | null;

  discount_amount?: string | null;

  original_price?: string | null;

  promo_price?: string | null;

  promo_code?: string | null;

  usage_limit?: number | null;

  usage_count?: number | null;

  service_slug?: string | null;

  cta_text?: string | null;

  cta_url?: string | null;

  conditions?: string | null;

  starts_at?: string | null;

  ends_at?: string | null;

  is_active?: boolean | null;

  sort_order?: number | null;

  created_at?: string | null;

  updated_at?: string | null;
}